qa/deepsea: split off deepsea_deploy task from deepsea task
deepsea task will just install deepsea
deepsea_deploy task will actually deploy a cluster
